File now parses.
I basically had to give up on all traces of polymorphism to make this
code work, and also have directives be static values initialized with
`Lazy.new`, not my favorite but alright.

I think this code can be cleaned up further by removing a lot of my
functional magic, it tends not to interact super well with lifetimes.